Fix inclusion of all libraries and dependencies when building with Eclipse.
The trick is that we needed to specify the openxc project dependency under the
"Android" settings for the project in Eclipse, not the regular build path.

Unfortunately there doesn't seem to be a way to use a path relative to some
common variable that all developers will have, so everyone will need to make
sure the openxc library is in the same path relative to this app.